refactor(api): split server.go (1685 → 7 files) by concern
Fourth worst-first file-size refactor. Function bodies relocated unchanged;
two small lint-driven cleanups on the moved lines.

  server.go              362   Server struct + NewServer + service accessors
  server_init.go         336   per-subsystem initialisation helpers
  server_routes.go       252   route table + per-module setup helpers
  server_middleware.go   149   security headers + CORS + body limit + recover
  server_spa.go          111   SPA static file fallback
  server_lifecycle.go    363   Start + HTTP/HTTPS/ACME + self-signed cert
  server_shutdown.go     197   Shutdown + data retention + link-state events

All under the 600-line warn budget.

Drive-by cleanups forced by --new-from-rev on relocated lines:

- Extracted jsonKeyInterface/jsonKeyTimestamp package-level constants
  alongside the existing statusResponse type in handlers_health_checks_
  settings.go; replaced the two relocated SSE-payload literals to silence
  goconst (the literals appear 20+ / 8+ times across the api package, but
  only the moved ones surfaced via --new-from-rev).
- Converted the recoverMiddleware panic log to its ErrorContext variant
  threading r.Context() for sloglint.
- Suppressed gosec G710 (open-redirect taint) on the HTTP→HTTPS redirect
  builder with inline justification: the redirect URL is constructed from
  server-controlled scheme + bare host + our own configured HTTPS port;
  user-supplied r.RequestURI is appended as path/query only.

Build, vet, lint (--new-from-rev=HEAD~1: 0 issues), and api tests pass.
